# packages/contracts/openapi/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Media Analysis Platform API",
    "version": "2.0.0",
    "description": "Single-user channel-aware target contract for API-owned media assets, collections, immutable selection snapshots, analysis runs, artifacts, diagnostics, channel accounts, channel surfaces, and idempotency."
  },
  "paths": {
...
    "/internal/v1/metadata-enrichment-jobs/{enrichment_id}/claim": {
      "post": {
        "operationId": "claimMetadataEnrichment",
        "summary": "Acquire a fenced metadata enrichment attempt and finite lease.",
        "parameters": [{"$ref": "#/components/parameters/EnrichmentId"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "../schemas/internal/metadata-enrichment-control.schema.json#/$defs/claimRequest"}}}},
        "responses": {"200": {"description": "Claim and canonical provider input.", "content": {"application/json": {"schema": {"$ref": "../schemas/internal/metadata-enrichment-control.schema.json#/$defs/claimResponse"}}}}, "409": {"$ref": "#/components/responses/ErrorEnvelope"}}
      }
    },
    "/internal/v1/metadata-enrichment-jobs/{enrichment_id}/progress": {
      "post": {
        "operationId": "recordMetadataEnrichmentProgress",
        "summary": "Record progress and heartbeat while preserving the claimed lease duration.",
        "parameters": [{"$ref": "#/components/parameters/EnrichmentId"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "../schemas/internal/metadata-enrichment-control.schema.json#/$defs/progressRequest"}}}},
        "responses": {"204": {"description": "Progress accepted."}, "409": {"$ref": "#/components/responses/ErrorEnvelope"}}
      }
    },
    "/internal/v1/metadata-enrichment-jobs/{enrichment_id}/finalize": {
      "post": {
        "operationId": "finalizeMetadataEnrichment",
        "summary": "Fenced finalize with bounded provider metadata or delayed retry.",
        "parameters": [{"$ref": "#/components/parameters/EnrichmentId"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "../schemas/internal/metadata-enrichment-control.schema.json#/$defs/finalizeRequest"}}}},
        "responses": {"200": {"description": "Updated metadata enrichment state.", "content": {"application/json": {"schema": {"$ref": "../schemas/http/metadata-enrichment.schema.json#/$defs/metadataEnrichmentResponse"}}}}, "409": {"$ref": "#/components/responses/ErrorEnvelope"}}
      }
    },
    "/internal/v1/metadata-enrichment-jobs/reclaim": {
      "post": {
        "operationId": "reclaimMetadataEnrichments",
        "summary": "Requeue expired metadata leases with backoff or fail exhausted jobs.",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "../schemas/internal/metadata-enrichment-control.schema.json#/$defs/reclaimRequest"}}}},
        "responses": {"200": {"description": "Reclaim summary.", "content": {"application/json": {"schema": {"$ref": "../schemas/internal/metadata-enrichment-control.schema.json#/$defs/reclaimResponse"}}}}}
      }
    },
...
  },
  "components": {
    "parameters": {
...
      "EnrichmentId": {
        "name": "enrichment_id",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "../schemas/common/enums.schema.json#/$defs/uuid"
        }
      },
...
    },
    "responses": {
      "ErrorEnvelope": {
        "description": "Validation, channel-scope, conflict, retention, or diagnostic error envelope.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "../schemas/common/error-envelope.schema.json"
            }
          }
        }
      }
    },
...
  }
}